requirements.txt
Django==4.1.2               # Python Web Framework
django-filter==22.1         # Filtering support.
django-model-utils==4.2.0   # Utilities for Django models.
djangorestframework==3.14.0 # API framework
Faker==14.2.1               # Generador de información aleatoria para los test.
Markdown==3.4.1             # Markdown support for the browsable API.
pytest-drf==1.1.3           # Pytest plugin for Django REST Framework.
pytest-django==4.5.2        # Integración con Django.
python-dateutil==2.8.2      # Utilidades para trabajar con fechas.
